Update build image for 2018 BioBuilds releases
Switched to CentOS 7 as the base for both the x86_64 and ppc64le images;
done to reduce maintenance headaches and to make the build environment
more uniform across the hardware platforms we support.

This also lets us make use of Docker's new multi-arch manifests, so the
build process for BioBuilds conda packages can now simply refer to the
"biobuilds/build" image, regardless of the target architecture.